Make the DoInquiryResponse capture more raw data from the SOAP request and allow re-creation from raw XML
The response from the DoInquiry eligiblity check now captures the raw XML and can be recreated from stored XML.

Traces are available as part of the DoInquiry eligibility response.  Trizetto adds a trace_id ("99Trizetto") and a trace_number to each request that can be provided to them later as part of a support request.  The payer may also add a trace.

Fill in some testing gaps

Minor doc updates